Move connector endpoints to dedicated file
